requirements.txt
streamlit==1.35.0
langchain==0.2.1
sentence-transformers==2.7.0
faiss-cpu==1.8.0
llama-cpp-python==0.2.57
sentence-splitter==1.4
